Started building Jest tests
